update vscode and windows terminal keybindings
Vscode:
> ctrl + alt + hjkl splits the current editor into another group (by the
direction of the hkjl keys)

Windows terminal:

> ctrl + alt + j duplicates the current pane
Which is mostly useless, but nice for when I'll ever need to make new
panes of a non-default profile
I doubt that will ever happen, but two cmd panes ig?
More likely, it's for when I'm going to be trying another distro

> ctrl + alt + k duplicates the current tab
Now this is a bit more useful, because it keeps the title of the current
tab.
If I use this while being on Main, I keep the title, and only need to
change the color (which I find annoying that it doesn't do by default)