^!d relaunches the terminal; !l opens last link
As well as some other new remaps for the vscode terminal, which I intend
to use on a project by project basis.
The permanent terminals I wanna have will be in the windows terminal,
and terminals I nee for the project while I'm working on it will be in
vscode.
Let's see if that works for me