git pushes don't get devnulled anymore
As I start working with more different projects, especially those that
are not mine, the idea of silencing the push makes less sense because
there're more chances it will fail.
And then I don't know why it did, so I try repushing to see the error
message, which is really backwards.
I'm fine with scrolling my terminal a bit more, and I still have an
abbreviation that lets me devnull things if I do want to do so