remove post command clear -x 's
Post output clears are a bad idea.
You don't have enough time to consider the output.
And sure, you can scroll back, but you're just playing catch at that
point.
Clear *before* command execution is a great idea, not viable to make for
every single command / abbreviation, but for the most used commands,
it's the way to go.
Clear after command execution is fine too if it's for a command that
doesn't give output, and is obvious in whether it did what it
needed to