vimium: add more link related operators
Now `F` also automatically moves you to the tab you opened,
`o` does what `F` used to do before: opens a link and doesn't switch to
the new tab.
`O` is now an alternative to `!f`, letting you choose multiple links to
open in new tabs.
The reason for it existing is that I may potentially figure out how to
*unmap* hotkeys, so I can go back to using `!f` for toggling dark mode.
`q` lets you download a link, possibly giving an ability to download
pictures and such.
`[` lets you pass the next key as if you're doing insert mode (the
opposite of <C-o>), which isn't going to be useful often, but helpful
for when it will be.
`D` and `U` are now the full page alternatives to `d` and `u`. Strange
that they aren't set by default.