vscode: set background color for git glyphs properly
The code tells you otherwise, sure, but that's kinda how it's supposed
to be done.
If we set background-image, then the width of the glyphs are still
customizable via vscode settings (and I would rather still keep that option).
Setting background-color (like a sane person) makes the glyph about twice as large. I've tried a lot of ways to resize it, but couldn't find a way to make it conform as easily as this hacky solution.