vscode: move breakpoint circles a bit to the right
Despite having the git glyphs all the way on the left, they're actually
still usable!
And they look exactly as correct as vscode makes them to be anyway (not very).
The benefit of this is that it's not my css' issue!
In my default font size, the breakpoint circle ends up in the middle
between git glyphs and line numbers.
This keeps being the case with editor +- zooming, however both git glyphs and breakpoint circle don't scale up / down as much as the other stuff, so they look kinda weird. Once again not my issue!