systemd-timers: all services wait for a kitty window to exist
Because as I was able to figure from SEVEN HOURS OF THIS SHIT, this is
the most consistent way to make sure the services have access to DISPLAY
and XAUTHORITY from the environment.
You can set DISPLAY in the service file, but my XAUTHORITY isn't static,
so I needed a way to check for something else and transitively expect I
have access to my graphical session.
I'm basically guaranteed to open kitty when I log in, and pretty
quickly.
Usually it gets created automatically, but when it doesn't, *I* would
because something fucked up and I need a terminal to fix it.