awesome: use `.visible` to toggle hide widgets
Turns out it's like a noticeable speed improvement over changing the
margins, for whatever reason.
Like, any programmer will agree that using `.visible` here should
probably be better, and I agree; but I'm surprised that the difference
is actually *noticeable*.
Usually it'd just be theory mind games to pick the best theoretical
option, regardless if you as the user are going to notice its bestness.