xremap: move filter and restore, remove set and write
The thought process behind this, is that I barely ever want to make sure
to overwrite something in a register, and if it's empty, the semantics
of "appending" to it are exactly the same as writing to it.
The real workflow is that first I'll check what it contains with `<+e`,
or now, truncate it *and* check at the same time with `<+r`, then once
it's correctly empty I can append to it. I still would've checked
regardless!
Similar case with the clipboard thing, I might want to possibly "collect"
what's in my clipboard (links, for example), I won't really keep on
`set`ing to a register that isn't empty. And then again, I'd check
what's in there anyway.
`<+x` for "cut", meaning copy and truncate is on the fence of being
useful, considering I don't remember the last time I used it (lol)
So it will be the first thing to get removed in the future, for a
potential "fuzzy search file" action. Not sure if I need it yet though.