astro: remove most treesitter text objects
Most of them either target strange places, or are too specific to be
useful.
Some of them can be easily replaced by a simple `cib` or a `cii`.
Some of them miss the point, like the comment one.
Targets one, individual comment, while I want to act on the entire
block.