astro: telescope grep_string mappings search for word under cursor
or if called from visual mode, for the selection.
They used to *only* search for the default register, which doesn't make
much sense.
If I have something in my default register, it's not much of
a hassle to paste it in the usual live_grep.
For something this potentially common as searching for the word under
your cursor, it makes no sense to require the step of first copying it.
And with the visual mode being an option, it's a good enough tradeoff
for a nicer default.