astro: change kitty blank series hotkeys
now there are no longer hotkeys for creating an os window, since I've
literally never used that lol.
there's a mapping in kitty that lets me move a window into something
else, like into another tab, into a new tab, or an os window.
so if the nvim mappings aren't expansive enough, I can always use kitty
hotkeys further to reach the state that I want (considering it's a very
rare state to want)
in the future, I still have the option of ,dg for creating an os window,
but for now I simply do not need it
the reason why I'm changing the hotkeys in the first place, is that ,t
and ,v are going to be used for buffer local and global search harps, a
harp-nvim feature I'm working on