astro: disable nvim-spider subword movement
The whole idea of the plugin is to ignore useless stops, so it's a
strange balance for it to also move by subwords. I don't remember the
last time I needed to move by subword, and it not being me trying to
just *change* over a subword, so I don't think I'm gonna be losing out
on much.